Add a bootstrap script to the create the Thingpedia database
Creating the Thingpedia database is more than just creating
the schema and adding the root & anonymous users: one needs
to import a minimal set of entities, schemas and devices to
ensure the clients (and web almond) can find the resources
they expect.
This is very verbose to write in pure SQL (and would not
cover things like icons), so add a script to do that.

Also add a test that runs the script on a clean database.
Very soon we'll fill newly set up database with test
data and use it to test Thingpedia API.
