+++ data/messaging.manifest.json
@@ -1,0 +1,5 @@
+{
+    "queries": {},
+    "actions": {},
+    "examples": []
+}
