+++ data/org.thingpedia.builtin.bluetooth.generic.manifest.json
@@ -1,0 +1,17 @@
+{
+    "thingpedia_name": "Generic Bluetooth Device",
+    "thingpedia_description": "The most generic of Bluetooth classes, it supports no operation at all. It is used if nothing more specific is found.",
+    "module_type": "org.thingpedia.builtin",
+    "params": {},
+    "category": "physical",
+    "subcategory": "home",
+    "types": [],
+    "child_types": [],
+    "auth": {
+        "type": "discovery",
+        "discoveryType": "bluetooth"
+    },
+    "queries": {},
+    "actions": {},
+    "examples": []
+}
