+++ data/org.thingpedia.builtin.matrix.manifest.json
@@ -1,0 +1,18 @@
+{
+    "thingpedia_name": "Matrix",
+    "thingpedia_description": "The Matrix.org federated chat service",
+    "module_type": "org.thingpedia.builtin",
+    "params": {},
+    "category": "online",
+    "subcategory": "social-network",
+    "types": [
+        "messaging"
+    ],
+    "child_types": [],
+    "auth": {
+        "type": "interactive"
+    },
+    "queries": {},
+    "actions": {},
+    "examples": []
+}
