+++ data/org.thingpedia.builtin.test.manifest.json
@@ -1,0 +1,103 @@
+{
+    "thingpedia_name": "Test Device",
+    "thingpedia_description": "Test Almond in various ways",
+    "module_type": "org.thingpedia.builtin",
+    "params": {},
+    "category": "system",
+    "subcategory": "service",
+    "types": [],
+    "child_types": [],
+    "auth": {
+        "type": "builtin"
+    },
+    "queries": {
+        "get_data": {
+            "args": [
+                {
+                    "name": "size",
+                    "type": "Measure(byte)",
+                    "question": "How much fake data do you want?",
+                    "is_input": true,
+                    "required": true
+                },
+                {
+                    "name": "count",
+                    "type": "Number",
+                    "question": "",
+                    "is_input": true,
+                    "required": false
+                },
+                {
+                    "name": "data",
+                    "type": "String",
+                    "question": "",
+                    "is_input": false,
+                    "required": false
+                }
+            ],
+            "doc": "generate `size` amount of fake data",
+            "confirmation": "generate $size of fake data",
+            "confirmation_remote": "generate $size of fake data on $__person's Almond",
+            "canonical": "get data on test",
+            "formatted": [
+                {
+                    "type": "text",
+                    "text": "${data}"
+                }
+            ],
+            "poll_interval": 0,
+            "is_list": true
+        },
+        "dup_data": {
+            "args": [
+                {
+                    "name": "data_in",
+                    "type": "String",
+                    "question": "What data do you want to duplicate?",
+                    "is_input": true,
+                    "required": true
+                },
+                {
+                    "name": "data_out",
+                    "type": "String",
+                    "question": "",
+                    "is_input": false,
+                    "required": false
+                }
+            ],
+            "doc": "duplicate the data (concatenate two copies); this is a simple deterministic get that depends on the input and is used to test param passing into a get",
+            "confirmation": "duplicate ${data_in} data",
+            "confirmation_remote": "",
+            "canonical": "duplicate data on test",
+            "formatted": [
+                {
+                    "type": "text",
+                    "text": "${data_out}"
+                }
+            ],
+            "poll_interval": -1,
+            "is_list": false
+        }
+    },
+    "actions": {
+        "eat_data": {
+            "args": [
+                {
+                    "name": "data",
+                    "type": "String",
+                    "question": "What do you want me to consume?",
+                    "is_input": true,
+                    "required": true
+                }
+            ],
+            "doc": "consume some data, do nothing",
+            "confirmation": "consume $data",
+            "confirmation_remote": "consume $data on $__person's Almond",
+            "canonical": "eat data on test",
+            "formatted": [],
+            "poll_interval": -1,
+            "is_list": false
+        }
+    },
+    "examples": []
+}
