+++ data/org.thingpedia.builtin.thingengine.remote.manifest.json
@@ -1,0 +1,16 @@
+{
+    "thingpedia_name": "Remote ThingSystem",
+    "thingpedia_description": "A proxy device for a ThingSystem owned by a different user. This device is created and managed automatically by the system.",
+    "module_type": "org.thingpedia.builtin",
+    "params": {},
+    "category": "system",
+    "subcategory": "service",
+    "types": [],
+    "child_types": [],
+    "auth": {
+        "type": "builtin"
+    },
+    "actions": {},
+    "queries": {},
+    "examples": []
+}
