+++ routes/thingpedia_upload.js
@@ -14,27 +14,16 @@ const express = require('express');
 const fs = require('fs');
 const multer = require('multer');
 const csurf = require('csurf');
-const JSZip = require('jszip');
-const ThingTalk = require('thingtalk');
 
 const db = require('../util/db');
 const code_storage = require('../util/code_storage');
 const model = require('../model/device');
-const schema = require('../model/schema');
 const user = require('../util/user');
-const exampleModel = require('../model/example');
-const entityModel = require('../model/entity');
-const Validation = require('../util/validation');
-const ManifestToSchema = require('../util/manifest_to_schema');
-const TrainingServer = require('../util/training_server');
-const tokenizer = require('../util/tokenize');
-const graphics = require('../almond/graphics');
 const platform = require('../util/platform');
-const colorScheme = require('../util/color_scheme');
+const TrainingServer = require('../util/training_server');
+const Importer = require('../util/import_device');
 
 const EngineManager = require('../almond/enginemanagerclient');
-
-const PARAM_REGEX = /\$(?:([a-zA-Z0-9_]+(?![a-zA-Z0-9_]))|{([a-zA-Z0-9_]+)(?::([a-zA-Z0-9_]+))?})/;
 
 var router = express.Router();
 
@@ -62,258 +51,6 @@ router.get('/create', user.redirectLogIn, user.requireDeveloper(), (req, res) =>
                                                      create: true });
 });
 
-function schemaCompatible(s1, s2) {
-    return s1.length >= s2.length &&
-        s2.every((t, i) => {
-            if (t === s1[i]) return true;
-            var t1 = ThingTalk.Type.fromString(t);
-            var t2 = ThingTalk.Type.fromString(s1[i]);
-            return t1.equals(t2);
-        });
-}
-
-function validateSchema(dbClient, types, ast, req) {
-    if (types.length === 0)
-        return Promise.resolve();
-
-    return schema.getTypesAndNamesByKinds(dbClient, types, req.user.developer_org).then((rows) => {
-        if (rows.length < types.length)
-            throw new Error(req._("Invalid device types %s").format(types));
-
-        function validate(where, what, against, type) {
-            for (var name in against) {
-                if (!(name in where))
-                    throw new Error(req._("Type %s requires %s %s").format(type, what, name));
-                var types = where[name].args.map((a) => a.type);
-                if (!schemaCompatible(types, against[name].types))
-                    throw new Error(req._("Schema for %s is not compatible with type %s").format(name, type));
-            }
-        }
-
-        for (let row of rows) {
-            validate(ast.actions, 'action', row.actions, row.kind);
-            validate(ast.queries, 'query', row.queries, row.kind);
-        }
-    });
-}
-
-const CATEGORIES = new Set(['physical','data','online','system']);
-const CATEGORY_TYPES = new Set(['data-source', 'online-account', 'thingengine-system']);
-const SUBCATEGORIES = new Set(['service','media','social-network','communication','home','health','data-management']);
-const ALLOWED_MODULE_TYPES = new Set(['org.thingpedia.v2', 'org.thingpedia.v1', 'org.thingpedia.rss', 'org.thingpedia.rest_json', 'org.thingpedia.builtin', 'org.thingpedia.generic_rest.v1', 'org.thingpedia.embedded']);
-const JAVASCRIPT_MODULE_TYPES = new Set(['org.thingpedia.v1', 'org.thingpedia.v2', 'org.thingpedia.builtin']);
-const AUTH_TYPES = new Set(['none','oauth2','basic','builtin','discovery','interactive']);
-
-function validateDevice(dbClient, req) {
-    var name = req.body.name;
-    var description = req.body.description;
-    var code = req.body.code;
-    var kind = req.body.primary_kind;
-
-    if (!name || !description || !code || !kind)
-        throw new Error(req._("Not all required fields were present"));
-
-    var ast = JSON.parse(code);
-    if (!ast.module_type || !ALLOWED_MODULE_TYPES.has(ast.module_type))
-        throw new Error(req._("Invalid module type"));
-    const fullcode = !JAVASCRIPT_MODULE_TYPES.has(ast.module_type);
-
-    if (!ast.params)
-        ast.params = {};
-    if (!ast.types)
-        ast.types = [];
-    if (!ast.child_types)
-        ast.child_types = [];
-    if (!ast.auth)
-        ast.auth = {"type":"none"};
-    if (ast.module_type === 'org.thingpedia.embedded') {
-        ast.auth.type = 'builtin';
-        ast.params = {};
-    }
-    if (!ast.auth.type || !AUTH_TYPES.has(ast.auth.type))
-        throw new Error(req._("Invalid authentication type"));
-    if (ast.auth.type === 'basic' && (!ast.params.username || !ast.params.password))
-        throw new Error(req._("Username and password must be declared for basic authentication"));
-    if (!CATEGORIES.has(ast.category))
-        throw new Error(req._("Invalid category %s").format(ast.category));
-    if (!SUBCATEGORIES.has(ast.subcategory))
-        throw new Error(req._("Invalid device domain %s").format(ast.subcategory));
-    //if (ast.auth.type === 'oauth2' && !ast.auth.client_id && !ast.auth.client_secret)
-    //    throw new Error(req._("Client ID and Client Secret must be provided for OAuth 2 authentication"));
-    ast.types = ast.types.filter((t) => {
-        return !CATEGORY_TYPES.has(t) && !SUBCATEGORIES.has(t);
-    });
-    ast.child_types.forEach((t) => {
-        if (CATEGORY_TYPES.has(t) || SUBCATEGORIES.has(t))
-            throw new Error(req._("Cannot specify category %s as child type").format(t));
-    });
-    if (ast['global-name'])
-        throw new Error(req._("Global names are obsolete, remove them"));
-
-    return Promise.resolve().then(() => {
-        return Validation.validateAllInvocations(kind, ast);
-    }).then((entities) => {
-        return entityModel.checkAllExist(dbClient, Array.from(entities));
-    }).then(() => {
-        if (fullcode && ast.module_type !== 'org.thingpedia.embedded') {
-            if (!ast.name)
-                ast.name = name;
-            if (!ast.description)
-                ast.description = description;
-            for (let name in ast.triggers) {
-                if (!ast.triggers[name].url)
-                    throw new Error(req._("Missing trigger url for %s").format(name));
-            }
-            for (let name in ast.actions) {
-                if (!ast.actions[name].url)
-                    throw new Error(req._("Missing action url for %s").format(name));
-            }
-            for (let name in ast.queries) {
-                if (!ast.queries[name].url)
-                    throw new Error(req._("Missing query url for %s").format(name));
-            }
-        }
-
-        return validateSchema(dbClient, ast.types, ast, req);
-    }).then(() => ast);
-}
-
-function ensurePrimarySchema(dbClient, name, kind, ast, req, approve) {
-    const metas = ManifestToSchema.toSchema(ast);
-
-    return schema.getByKind(dbClient, kind).then((existing) => {
-        if (existing.owner !== req.user.developer_org &&
-            req.user.developer_status < user.DeveloperStatus.ADMIN)
-            throw new Error(req._("Not Authorized"));
-
-        var obj = {
-            kind_canonical: tokenizer.tokenize(name).join(' '),
-            developer_version: existing.developer_version + 1
-        };
-        if (req.user.developer_status >= user.DeveloperStatus.TRUSTED_DEVELOPER && approve)
-            obj.approved_version = obj.developer_version;
-
-        return schema.update(dbClient, existing.id, existing.kind, obj, metas);
-    }, (e) => {
-        var obj = {
-            kind: kind,
-            kind_canonical: tokenizer.tokenize(name).join(' '),
-            kind_type: 'primary',
-            owner: req.user.developer_org
-        };
-        if (req.user.developer_status < user.DeveloperStatus.TRUSTED_DEVELOPER || !approve) {
-            obj.approved_version = null;
-            obj.developer_version = 0;
-        } else {
-            obj.approved_version = 0;
-            obj.developer_version = 0;
-        }
-        return schema.create(dbClient, obj, metas);
-    }).then((schema) => {
-        return ensureExamples(dbClient, schema.id, ast);
-    });
-}
-
-function ensureExamples(dbClient, schemaId, ast) {
-    return exampleModel.deleteBySchema(dbClient, schemaId, 'en').then(() => {
-        return Validation.tokenizeAllExamples('en', ast.examples);
-    }).then((examples) => {
-        return exampleModel.createMany(dbClient, examples.map((ex) => {
-            return ({
-                schema_id: schemaId,
-                utterance: ex.utterance,
-                preprocessed: ex.preprocessed,
-                target_code: ex.program,
-                target_json: '', // FIXME
-                type: 'thingpedia',
-                language: 'en',
-                is_base: 1
-            });
-        }));
-    });
-}
-
-function uploadZipFile(req, obj, ast, stream) {
-    var zipFile = new JSZip();
-
-    return Promise.resolve().then(() => {
-        // unfortunately JSZip only loads from memory, so we need to load the entire file
-        // at once
-        // this is somewhat a problem, because the file can be up to 30-50MB in size
-        // we just hope the GC will get rid of the buffer quickly
-
-        var buffers = [];
-        var length = 0;
-        return new Promise((callback, errback) => {
-            stream.on('data', (buffer) => {
-                buffers.push(buffer);
-                length += buffer.length;
-            });
-            stream.on('end', () => {
-                callback(Buffer.concat(buffers, length));
-            });
-            stream.on('error', errback);
-        });
-    }).then((buffer) => {
-        return zipFile.loadAsync(buffer, { checkCRC32: false });
-    }).then(() => {
-        var packageJson = zipFile.file('package.json');
-        if (!packageJson)
-            throw new Error(req._("package.json missing from device zip file"));
-
-        return packageJson.async('string');
-    }).then((text) => {
-        try {
-            var parsed = JSON.parse(text);
-        } catch(e) {
-            throw new Error("Invalid package.json: SyntaxError at line " + e.lineNumber + ": " + e.message);
-        }
-        if (!parsed.name || !parsed.main)
-            throw new Error(req._("Invalid package.json (missing name or main)"));
-
-        parsed['thingpedia-version'] = obj.developer_version;
-
-        zipFile.file('package.json', JSON.stringify(parsed));
-
-        return code_storage.storeZipFile(zipFile.generateNodeStream({ compression: 'DEFLATE',
-                                                                      type: 'nodebuffer',
-                                                                      platform: 'UNIX'}),
-                                         obj.primary_kind, obj.developer_version);
-    }).catch((e) => {
-        console.error('Failed to upload zip file to S3: ' + e);
-        console.error(e.stack);
-        throw e;
-    });
-}
-
-function isJavaScript(file) {
-    return file.mimetype === 'application/javascript' ||
-        file.mimetype === 'text/javascript' ||
-        (file.originalname && file.originalname.endsWith('.js'));
-}
-
-function uploadJavaScript(req, obj, ast, stream) {
-    var zipFile = new JSZip();
-
-    return Promise.resolve().then(() => {
-        zipFile.file('package.json', JSON.stringify({
-            name: obj.primary_kind,
-            author: req.user.username + '@thingpedia.stanford.edu',
-            main: 'index.js',
-            'thingpedia-version': obj.developer_version
-        }));
-        zipFile.file('index.js', stream);
-        return code_storage.storeZipFile(zipFile.generateNodeStream({ compression: 'DEFLATE',
-                                                                      type: 'nodebuffer',
-                                                                      platform: 'UNIX'}),
-                                         obj.primary_kind, obj.developer_version);
-    }).catch((e) => {
-        console.error('Failed to upload zip file to S3: ' + e);
-        console.error(e.stack);
-        throw e;
-    });
-}
-
 function tryUpdateDevice(primaryKind, userId) {
     // do the update asynchronously - if the update fails, the user will
     // have another chance from the status page
@@ -324,6 +61,12 @@ function tryUpdateDevice(primaryKind, userId) {
     });
 
     return Promise.resolve();
+}
+
+function isJavaScript(file) {
+    return file.mimetype === 'application/javascript' ||
+        file.mimetype === 'text/javascript' ||
+        (file.originalname && file.originalname.endsWith('.js'));
 }
 
 function doCreateOrUpdate(id, create, req, res) {
@@ -340,7 +83,7 @@ function doCreateOrUpdate(id, create, req, res) {
             return Promise.resolve().then(() => {
                 if (create && (!req.files.icon || !req.files.icon.length))
                     throw new Error(req._("An icon must be specified for new devices"));
-                return validateDevice(dbClient, req);
+                return Importer.validateDevice(dbClient, req);
             }).catch((e) => {
                 res.render('thingpedia_device_create_or_edit', { page_title:
                                                                  (create ?
@@ -359,7 +102,7 @@ function doCreateOrUpdate(id, create, req, res) {
                 if (ast === null)
                     return null;
 
-                return ensurePrimarySchema(dbClient, name, kind, ast, req, approve).then(() => ast);
+                return Importer.ensurePrimarySchema(dbClient, name, kind, ast, req, approve).then(() => ast);
             }).then((ast) => {
                 if (ast === null)
                     return null;
@@ -367,7 +110,7 @@ function doCreateOrUpdate(id, create, req, res) {
                 var extraKinds = ast.types;
                 var extraChildKinds = ast.child_types;
 
-                var fullcode = !JAVASCRIPT_MODULE_TYPES.has(ast.module_type);
+                var fullcode = Importer.isFullCode(ast.module_type);
 
                 var obj = {
                     primary_kind: kind,
@@ -415,25 +158,30 @@ function doCreateOrUpdate(id, create, req, res) {
                             });
                     });
                 }
-            }).then((obj) => {
+            }).then(async (obj) => {
                 if (obj === null)
                     return null;
 
                 if (obj.fullcode || gAst.module_type === 'org.thingpedia.builtin')
                     return obj.primary_kind;
 
-                if (req.files && req.files.zipfile && req.files.zipfile.length &&
-                    isJavaScript(req.files.zipfile[0]))
-                    return uploadJavaScript(req, obj, gAst, fs.createReadStream(req.files.zipfile[0].path)).then(() => obj.primary_kind);
+                const zipFile = req.files && req.files.zipfile && req.files.zipfile.length ?
+                    req.files.zipfile[0] : null;
 
                 let stream;
-                if (req.files && req.files.zipfile && req.files.zipfile.length)
-                    stream = fs.createReadStream(req.files.zipfile[0].path);
+                if (zipFile !== null)
+                    stream = fs.createReadStream(zipFile.path);
                 else if (obj.old_version !== null)
                     stream = code_storage.downloadZipFile(obj.primary_kind, obj.old_version);
                 else
                     throw new Error(req._("Invalid zip file"));
-                return uploadZipFile(req, obj, gAst, stream).then(() => obj.primary_kind);
+
+                if (zipFile && isJavaScript(zipFile))
+                    await Importer.uploadJavaScript(req, obj, gAst, stream);
+                else
+                    await Importer.uploadZipFile(req, obj, gAst, stream);
+
+                return obj.primary_kind;
             }).then((done) => {
                 if (!done)
                     return done;
@@ -441,22 +189,7 @@ function doCreateOrUpdate(id, create, req, res) {
                 if (req.files.icon && req.files.icon.length) {
                     // upload the icon asynchronously to avoid blocking the request
                     setTimeout(() => {
-                        Q.try(() => {
-                            var image = graphics.createImageFromPath(req.files.icon[0].path);
-                            image.resizeFit(512, 512);
-                            return image.stream('png');
-                        }).then(([stdout, stderr]) => {
-                            console.error('here');
-                            return code_storage.storeIcon(stdout, done);
-                        }).then(() => {
-                            console.error('there');
-                            return colorScheme(req.files.icon[0].path, kind);
-                        }).catch ((e) => {
-                            console.error('Failed to upload icon to S3: ' + e);
-                            console.error(e.stack);
-                        }).finally(() => {
-                            return Q.nfcall(fs.unlink, req.files.icon[0].path);
-                        });
+                        Importer.uploadIcon(done, req.files.icon[0].path);
                     }, 0);
                 }
 
@@ -494,109 +227,6 @@ router.post('/create', user.requireLogIn, user.requireDeveloper(), (req, res) =>
     doCreateOrUpdate(undefined, true, req, res);
 });
 
-function legacyCreateExample(utterance, kind, function_name, function_type, function_obj) {
-    let inargmap = {};
-    let outargmap = {};
-    for (let arg of function_obj.args) {
-        if (arg.is_input)
-            inargmap[arg.name] = arg.type;
-        else
-            outargmap[arg.name] = arg.type;
-    }
-
-    let regexp = new RegExp(PARAM_REGEX, 'g');
-
-    let in_args = '';
-    let filter = '';
-    let arg_decl = '';
-    let any_arg = false;
-    let any_in_arg = false;
-    let any_out_arg = false;
-
-    let match = regexp.exec(utterance);
-    while (match !== null) {
-        let [, param1, param2,] = match;
-        let param = param1 || param2;
-
-        if (param in inargmap) {
-            let type = inargmap[param];
-            if (any_in_arg)
-                in_args += ', ';
-            if (any_arg)
-                arg_decl += ', ';
-            in_args += `${param}=p_${param}`;
-            arg_decl += `p_${param} :${type}`;
-            any_in_arg = true;
-            any_arg = true;
-        } else {
-            let type = outargmap[param];
-            if (any_out_arg)
-                filter += ' && ';
-            if (any_arg)
-                arg_decl += ', ';
-            filter += `${param} == p_${param}`;
-            arg_decl += `p_${param} :${type}`;
-            any_out_arg = true;
-            any_arg = true;
-        }
-
-        match = regexp.exec(utterance);
-    }
-
-    let result = `@${kind}.${function_name}(${in_args})`;
-    if (filter !== '')
-        result += `, ${filter}`;
-    if (function_type === 'triggers')
-        result = `let stream x := \\(${arg_decl}) -> monitor ${result};`;
-    else if (function_type === 'queries')
-        result = `let table x := \\(${arg_decl}) -> ${result};`;
-    else
-        result = `let action x := \\(${arg_decl}) -> ${result};`;
-    return { utterance: utterance, program: result };
-}
-
-function migrateManifest(code, device) {
-    let ast = JSON.parse(code);
-
-    delete ast.global_name;
-    delete ast['global-name'];
-    ast.category = device.category;
-    ast.subcategory = device.subcategory;
-
-    ast.types = (ast.types || []).filter((t) => {
-        if (CATEGORY_TYPES.has(t) || SUBCATEGORIES.has(t))
-            return false;
-        return true;
-    });
-    ast.child_types = ast.child_types || [];
-
-    if (!ast.examples) {
-        ast.examples = [];
-
-        for (let function_type of ['triggers','queries','actions']) {
-            for (let function_name in ast[function_type]) {
-                let function_obj = ast[function_type][function_name];
-
-                for (let example of (function_obj.examples || []))
-                    ast.examples.push(legacyCreateExample(example, device.primary_kind, function_name, function_type, function_obj));
-                delete function_obj.examples;
-            }
-        }
-    }
-
-    for (let function_type of ['triggers','queries']) {
-        for (let function_name in ast[function_type]) {
-            let function_obj = ast[function_type][function_name];
-
-            if (!('poll_interval' in function_obj))
-                function_obj.poll_interval = function_obj['poll-interval'] || -1;
-            delete function_obj['poll-interval'];
-        }
-    }
-
-    return JSON.stringify(ast);
-}
-
 router.get('/update/:id', user.redirectLogIn, user.requireDeveloper(), (req, res, next) => {
     Promise.resolve().then(() => {
         return db.withClient((dbClient) => {
@@ -606,7 +236,7 @@ router.get('/update/:id', user.redirectLogIn, user.requireDeveloper(), (req, res
                     throw new Error(req._("Not Authorized"));
 
                 return model.getCodeByVersion(dbClient, req.params.id, d.developer_version).then((row) => {
-                    d.code = migrateManifest(row.code, d);
+                    d.code = Importer.migrateManifest(row.code, d);
                     return d;
                 });
             }).then((d) => {
@@ -643,7 +273,7 @@ router.get('/example/:id', (req, res, next) => {
         return db.withClient((dbClient) => {
             return model.get(dbClient, req.params.id).then((d) => {
                 return model.getCodeByVersion(dbClient, req.params.id, d.developer_version).then((row) => {
-                    d.code = migrateManifest(row.code, d);
+                    d.code = Importer.migrateManifest(row.code, d);
                     let ast = JSON.parse(d.code);
                     if ('client_id' in ast.auth)
                         ast.auth.client_id = '*** your-own-client-id ***';
